@@ -3,6 +3,8 @@ import { useNavigate } from 'react-router-dom';
 
 const Header = ({ onLogout }) => {
     const navigate = useNavigate();
+    const role = localStorage.getItem('role');
+    const isAdmin = role === 'ADMIN';
 
     return (
         <header className="app-header">
@@ -10,12 +12,16 @@ const Header = ({ onLogout }) => {
                 NPK School
             </h2>
             <div className="nav-buttons">
-                <button onClick={() => navigate('/students')} className="nav-button">
-                    Quản lý học sinh
-                </button>
-                <button onClick={() => navigate('/teachers')} className="nav-button">
-                    Quản lý giáo viên
-                </button>
+                {isAdmin && (
+                    <>
+                        <button onClick={() => navigate('/students')} className="nav-button">
+                            Quản lý học sinh
+                        </button>
+                        <button onClick={() => navigate('/teachers')} className="nav-button">
+                            Quản lý giáo viên
+                        </button>
+                    </>
+                )}
                 <button onClick={onLogout} className="logout-button">
                     Đăng xuất
                 </button>
