@@ -1,9 +1,10 @@
 import axios from 'axios';
 
 const studentApi = axios.create({
-    baseURL: 'http://localhost:8083/api/student',
+    baseURL: 'http://localhost:8083/api/student'
 });
 
+// Add a request interceptor to add the token to all requests
 studentApi.interceptors.request.use(
     (config) => {
         const token = localStorage.getItem('token');
@@ -12,7 +13,9 @@ studentApi.interceptors.request.use(
         }
         return config;
     },
-    (error) => Promise.reject(error)
+    (error) => {
+        return Promise.reject(error);
+    }
 );
 
 export default studentApi;
