@@ -1,9 +1,11 @@
 package com.sagroup.studentservice.domains;
 
 import javax.persistence.*;
+import javax.validation.constraints.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import com.sagroup.studentservice.constants.BehaviorGrade;
 
 @Entity
 @Table(name = "students")
@@ -17,24 +19,57 @@ public class Student {
     private Long studentNumber;
 
     @Column(nullable = false)
+    @NotBlank(message = "Họ không được để trống")
+    @Pattern(regexp = "^[A-ZÀÁẠẢÃÂẦẤẬẨẪĂẰẮẶẲẴÈÉẸẺẼÊỀẾỆỂỄÌÍỊỈĨÒÓỌỎÕÔỒỐỘỔỖƠỜỚỢỞỠÙÚỤỦŨƯỪỨỰỬỮỲÝỴỶỸĐ][a-zàáạảãâầấậẩẫăằắặẳẵèéẹẻẽêềếệểễìíịỉĩòóọỏõôồốộổỗơờớợởỡùúụủũưừứựửữỳýỵỷỹđ]*(?:[ ][A-ZÀÁẠẢÃÂẦẤẬẨẪĂẰẮẶẲẴÈÉẸẺẼÊỀẾỆỂỄÌÍỊỈĨÒÓỌỎÕÔỒỐỘỔỖƠỜỚỢỞỠÙÚỤỦŨƯỪỨỰỬỮỲÝỴỶỸĐ][a-zàáạảãâầấậẩẫăằắặẳẵèéẹẻẽêềếệểễìíịỉĩòóọỏõôồốộổỗơờớợởỡùúụủũưừứựửữỳýỵỷỹđ]*)*$", 
+            message = "Họ phải viết hoa chữ cái đầu")
     private String firstName;
 
     @Column(nullable = false)
+    @NotBlank(message = "Tên không được để trống")
+    @Pattern(regexp = "^[A-ZÀÁẠẢÃÂẦẤẬẨẪĂẰẮẶẲẴÈÉẸẺẼÊỀẾỆỂỄÌÍỊỈĨÒÓỌỎÕÔỒỐỘỔỖƠỜỚỢỞỠÙÚỤỦŨƯỪỨỰỬỮỲÝỴỶỸĐ][a-zàáạảãâầấậẩẫăằắặẳẵèéẹẻẽêềếệểễìíịỉĩòóọỏõôồốộổỗơờớợởỡùúụủũưừứựửữỳýỵỷỹđ]*$", 
+            message = "Tên phải viết hoa chữ cái đầu")
     private String lastName;
 
     @Column(nullable = false, unique = true)
-    private String email; // Thêm trường email
+    @NotBlank(message = "Email không được để trống")
+    @Email(message = "Email không hợp lệ")
+    @Pattern(regexp = "^[a-zA-Z0-9._%+-]+@gmail\\.com$", message = "Email phải là địa chỉ Gmail")
+    private String email;
 
-    private String phone;  // Thêm trường phone
-    private String address; // Thêm trường address
+    @Pattern(regexp = "^(0|\\+84)(3|5|7|8|9)[0-9]{8}$", message = "Số điện thoại không hợp lệ")
+    private String phone;
 
+    private String address;
+
+    @DecimalMin(value = "0.0", message = "Điểm không được nhỏ hơn 0")
+    @DecimalMax(value = "10.0", message = "Điểm không được lớn hơn 10")
     private Double mathScore;
+
+    @DecimalMin(value = "0.0", message = "Điểm không được nhỏ hơn 0")
+    @DecimalMax(value = "10.0", message = "Điểm không được lớn hơn 10")
     private Double physicsScore;
+
+    @DecimalMin(value = "0.0", message = "Điểm không được nhỏ hơn 0")
+    @DecimalMax(value = "10.0", message = "Điểm không được lớn hơn 10")
     private Double chemistryScore;
+
+    @DecimalMin(value = "0.0", message = "Điểm không được nhỏ hơn 0")
+    @DecimalMax(value = "10.0", message = "Điểm không được lớn hơn 10")
     private Double literatureScore;
+
+    @DecimalMin(value = "0.0", message = "Điểm không được nhỏ hơn 0")
+    @DecimalMax(value = "10.0", message = "Điểm không được lớn hơn 10")
     private Double englishScore;
 
     private String academicPerformance;
-    private String behaviorScore;
+
+    @NotNull(message = "Hạnh kiểm không được để trống")
+    @Enumerated(EnumType.STRING)
+    private BehaviorGrade behaviorScore;
+
+    @NotBlank(message = "Lớp không được để trống")
     private String className;
+    
+    @Transient
+    private Double averageScore;
 }
