@@ -1,12 +1,32 @@
 package com.sagroup.studentservice.dto;
+
 import lombok.Data;
+import javax.validation.constraints.*;
+import com.sagroup.studentservice.constants.BehaviorGrade;
+
 @Data
 public class ScoreUpdateDTO {
+    @DecimalMin(value = "0.0", message = "Điểm không được nhỏ hơn 0")
+    @DecimalMax(value = "10.0", message = "Điểm không được lớn hơn 10")
     private Double mathScore;
+
+    @DecimalMin(value = "0.0", message = "Điểm không được nhỏ hơn 0")
+    @DecimalMax(value = "10.0", message = "Điểm không được lớn hơn 10")
     private Double physicsScore;
+
+    @DecimalMin(value = "0.0", message = "Điểm không được nhỏ hơn 0")
+    @DecimalMax(value = "10.0", message = "Điểm không được lớn hơn 10")
     private Double chemistryScore;
+
+    @DecimalMin(value = "0.0", message = "Điểm không được nhỏ hơn 0")
+    @DecimalMax(value = "10.0", message = "Điểm không được lớn hơn 10")
     private Double literatureScore;
+
+    @DecimalMin(value = "0.0", message = "Điểm không được nhỏ hơn 0")
+    @DecimalMax(value = "10.0", message = "Điểm không được lớn hơn 10")
     private Double englishScore;
-    private String behaviorScore;
+
+    @NotNull(message = "Hạnh kiểm không được để trống")
+    private BehaviorGrade behaviorScore;
 }
 
