@@ -3,12 +3,15 @@ package com.sagroup.userservice.controller;
 import com.sagroup.userservice.dtos.NewAppUserDto;
 import com.sagroup.userservice.dtos.NewUserRequest;
 import com.sagroup.userservice.dtos.ResetPasswordRequest;
+import com.sagroup.userservice.dtos.ChangePasswordRequest;
 import com.sagroup.userservice.entity.NewAppUser;
 import com.sagroup.userservice.service.NewAppUserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
+import com.sagroup.userservice.repository.NewAppUserRepository;
 
 import java.util.List;
 
@@ -18,6 +21,8 @@ import java.util.List;
 public class NewAppUserController {
 
     private NewAppUserService userService;
+    @Autowired
+    private NewAppUserRepository userRepository;
 
     @GetMapping("/view")
     public List<NewAppUser> viewAll(){
@@ -42,6 +47,17 @@ public class NewAppUserController {
         return ResponseEntity.ok("Password reset successfully.");
     }
 
+    @PostMapping("/auth/change-password")
+    public ResponseEntity<?> changePassword(@RequestBody ChangePasswordRequest request, Authentication authentication) {
+        try {
+            String username = authentication.getName();
+            userService.changePassword(username, request.getCurrentPassword(), request.getNewPassword());
+            return ResponseEntity.ok("Password changed successfully");
+        } catch (RuntimeException e) {
+            return ResponseEntity.badRequest().body(e.getMessage());
+        }
+    }
+
     @PutMapping
     public ResponseEntity<?> update(@RequestBody NewAppUserDto userDto){
         NewAppUser user = userService.update(userDto);
@@ -62,11 +78,24 @@ public class NewAppUserController {
     public ResponseEntity<?> createAccount(@RequestBody NewUserRequest request) {
         NewAppUserDto dto = new NewAppUserDto();
         dto.setEmail(request.getEmail());
-        dto.setUsername(request.getEmail());
+        dto.setUsername(request.getEmail()); // hoặc tách riêng username
         dto.setRole(request.getRole());
-        dto.setPassword("123456");
-        NewAppUser user = userService.save(dto);
-        return new ResponseEntity<>(user, HttpStatus.CREATED);
+
+        String rawPassword = (request.getPassword() == null || request.getPassword().isBlank()) ? "123456" : request.getPassword();
+        dto.setPassword(rawPassword);
+
+        try {
+            NewAppUser user = userService.save(dto);
+            return new ResponseEntity<>(user, HttpStatus.CREATED);
+        } catch (RuntimeException e) {
+            return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
+        }
+    }
+
+    @GetMapping("/check-email")
+    public ResponseEntity<Boolean> checkEmailExists(@RequestParam String email) {
+        boolean exists = userRepository.existsByUsernameIgnoreCase(email);
+        return ResponseEntity.ok(exists);
     }
 
 }
